refactor:MyPage -> MyPageEdit
