+++ frontend/src/router/index.js
@@ -11,7 +11,7 @@ import RoutineCommunity from '../pages/RoutineCommunity/RoutineCommunity.vue';
 import MyRoutineEdit from '@/pages/MyRoutine/MyRoutineEdit.vue';
 import Login from '../pages/Login&SignUp/Login.vue'
 import SignUp from '@/pages/Login&SignUp/SignUp.vue';
-import MyPage from '@/pages/Mypage/MyPage.vue';
+import MyPageEdit from '@/pages/Mypage/MyPageEdit.vue';
 import Register from '@/pages/Login&SignUp/Register.vue';
 
 
@@ -80,9 +80,9 @@ const routes = [
     ],
   },
   {
-    path: '/mypage',
-    name: 'mypage',
-    component: MyPage,
+    path: '/myedit',
+    name: 'myedit',
+    component: MyPageEdit,
   },
 ]
